% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{accuracy}
\alias{accuracy}
\title{Calculate the accuracy of predictions}
\usage{
accuracy(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
accuracy value
}
\description{
Calculate the accuracy of predictions
}
\examples{
# Define predicted and true values
pred <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
truth <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Calculate accuracy
accuracy(pred, truth)

}
