% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add two numbers together}
\usage{
add(number1, number2)
}
\arguments{
\item{number1}{A numeric value representing the first number to be added.}

\item{number2}{A numeric value representing the second number to be added.}
}
\value{
A numeric value representing the sum of the two input numbers.
}
\description{
This function takes two numeric inputs and returns their sum.
}
\examples{
add(3, 4) # returns 7
add(-2, 5) # returns 3
add(0, 0) # returns 0
add(1.5, 2.5) # returns 4
}
