% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{adj_R2}
\alias{adj_R2}
\title{Calculate the adjusted R-squared value of predictions}
\usage{
adj_R2(pred, truth, n_p)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}

\item{n_p}{number of predictor variables}
}
\value{
adjusted R-squared value
}
\description{
Calculate the adjusted R-squared value of predictions
}
\examples{
# Define true and predicted values
truth <- c(3, -0.5, 2, 7)
pred <- c(2.5, 0.0, 2, 8)
n_p <- 3

# Calculate adjusted R-squared value
adj_R2(pred, truth, n_p)
}
