% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{classes}
\alias{classes}
\title{Get the classes of columns in a dataframe or the class of an object}
\usage{
classes(x)
}
\arguments{
\item{x}{a dataframe or an object}
}
\value{
a character vector of class names or a single class name
}
\description{
Get the classes of columns in a dataframe or the class of an object
}
\examples{
# Create a dataframe
df <- data.frame(a = 1:3, b = c("one", "two", "three"), stringsAsFactors = FALSE)

# Get the classes of columns in the dataframe
column_classes <- classes(df)

         a           b
      "integer" "character"
# Get the class of an object (a numeric vector)
num_class <- classes(1:3)
[1] "integer"

}
