% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{f1}
\alias{f1}
\title{Calculate the F1 score of predictions}
\usage{
f1(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
F1 score value
}
\description{
Calculate the F1 score of predictions
}
\examples{
# Define predicted and true values
pred <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
truth <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Calculate F1 score
f1(pred, truth)
}
