% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_max}
\alias{is_max}
\title{Check if an element in a vector is the maximum}
\usage{
is_max(x, na.rm = T)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{remove NA values (default = TRUE)}
}
\value{
a logical vector
}
\description{
Check if an element in a vector is the maximum
}
\examples{
# Example 1
is_max(c(1, 2, 3, 4, 5))
# Example 2
is_max(c(NA, 1, 2, 3, 4, 5))

}
