% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{is_min}
\alias{is_min}
\title{Check if an element in a vector is the minimum}
\usage{
is_min(x, na.rm = T)
}
\arguments{
\item{x}{a numeric vector}

\item{na.rm}{remove NA values (default = TRUE)}
}
\value{
a logical vector
}
\description{
Check if an element in a vector is the minimum
}
\examples{
# Example 1
is_min(c(1, 2, 3, 4, 5))
TRUE FALSE FALSE FALSE FALSE
# Example 2
is_min(c(NA, 1, 2, 3, 4, 5))
TRUE FALSE FALSE FALSE FALSE

}
