% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_chisq}
\alias{log_likelihood_chisq}
\title{Calculate the log-likelihood of chi-squared distribution}
\usage{
log_likelihood_chisq(x, df)
}
\arguments{
\item{x}{a numeric vector}

\item{df}{degrees of freedom}
}
\value{
log-likelihood value
}
\description{
Calculate the log-likelihood of chi-squared distribution
}
\examples{
# Generate a sample from a chi-squared distribution
set.seed(42)
x <- rchisq(100, df = 5)

# Calculate the log-likelihood for the correct chi-squared distribution
log_likelihood_chisq(x, df = 5)

# Calculate the log-likelihood for a different chi-squared distribution
log_likelihood_chisq(x, df = 10)
}
