% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_f}
\alias{log_likelihood_f}
\title{Calculate the log-likelihood of F distribution}
\usage{
log_likelihood_f(x, df1, df2)
}
\arguments{
\item{x}{a numeric vector}

\item{df1}{numerator degrees of freedom}

\item{df2}{denominator degrees of freedom}
}
\value{
log-likelihood value
}
\description{
Calculate the log-likelihood of F distribution
}
\examples{
# Generate a sample from an F distribution
set.seed(42)
x <- rf(100, df1 = 5, df2 = 10)

# Calculate the log-likelihood for the correct F distribution
log_likelihood_f(x, df1 = 5, df2 = 10)

# Calculate the log-likelihood for a different F distribution
log_likelihood_f(x, df1 = 10, df2 = 5)
}
