% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_norm}
\alias{log_likelihood_norm}
\title{Calculate the log-likelihood of normal distribution}
\usage{
log_likelihood_norm(x, mean, sd)
}
\arguments{
\item{x}{a numeric vector}

\item{mean}{mean of the normal distribution}

\item{sd}{standard deviation of the normal distribution}
}
\value{
log-likelihood value
}
\description{
Calculate the log-likelihood of normal distribution
}
\examples{
x <- rnorm(100, mean = 5, sd = 2)
mean_est <- mean(x)
sd_est <- sd(x)
log_likelihood_norm(x, mean_est, sd_est)

}
