% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_t}
\alias{log_likelihood_t}
\title{Calculate the log-likelihood of t distribution}
\usage{
log_likelihood_t(x, df)
}
\arguments{
\item{x}{a numeric vector}

\item{df}{degrees of freedom}
}
\value{
log-likelihood value
}
\description{
Calculate the log-likelihood of t distribution
}
\examples{
# Generate a sample from a t distribution
set.seed(42)
x <- rt(100, df = 5)

# Calculate the log-likelihood for the correct t distribution
log_likelihood_t(x, df = 5)

# Calculate the log-likelihood for a different t distribution
log_likelihood_t(x, df = 10)

}
