% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{log_likelihood_unif}
\alias{log_likelihood_unif}
\title{Calculate the log-likelihood of uniform distribution}
\usage{
log_likelihood_unif(x, min, max)
}
\arguments{
\item{x}{a numeric vector}

\item{min}{minimum value of the uniform distribution}

\item{max}{maximum value of the uniform distribution}
}
\value{
log-likelihood value
}
\description{
Calculate the log-likelihood of uniform distribution
}
\examples{
# Generate a sample from a uniform distribution
set.seed(42)
x <- runif(100, min = 2, max = 5)

# Calculate the log-likelihood for the correct uniform distribution
log_likelihood_unif(x, min = 2, max = 5)

# Calculate the log-likelihood for a different uniform distribution
log_likelihood_unif(x, min = 0, max = 10)

}
