% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{minimum_n_per_group}
\alias{minimum_n_per_group}
\title{Calculate the minimum number of samples per group for a t-test}
\usage{
minimum_n_per_group(d, power = 0.8)
}
\arguments{
\item{d}{effect size}

\item{power}{desired power (default = 0.8)}
}
\value{
minimum number of samples per group
}
\description{
Calculate the minimum number of samples per group for a t-test
}
\examples{
# Define effect size and desired power
d <- 0.5
power <- 0.9

# Calculate minimum number of samples per group
minimum_n_per_group(d, power)
}
