% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{r2}
\alias{r2}
\title{Calculate the R-squared value of predictions}
\usage{
r2(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
R-squared value
}
\description{
Calculate the R-squared value of predictions
}
\examples{
# Define true and predicted values
truth <- c(3, -0.5, 2, 7)
pred <- c(2.5, 0.0, 2, 8)

# Calculate R-squared value
r2(pred, truth)
}
