% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rando}
\alias{rando}
\title{Generate random samples from a vector or dataframe-like object}
\usage{
rando(x, n = 1, replace = T)
}
\arguments{
\item{x}{an atomic vector or dataframe-like object}

\item{n}{number of samples to draw (default = 1)}

\item{replace}{whether to sample with replacement (default = TRUE)}
}
\value{
a sample or a dataframe-like object
}
\description{
Generate random samples from a vector or dataframe-like object
}
\examples{
# Sampling from a vector
vec <- 1:10
rando(vec, n = 3, replace = FALSE)

# Sampling from a data frame
df <- data.frame(a = 1:5, b = 6:10)
rando(df, n = 2, replace = TRUE)

# Sampling from a matrix
mat <- matrix(1:9, nrow = 3, ncol = 3)
rando(mat, n = 2, replace = FALSE)
}
