% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{recall}
\alias{recall}
\title{Calculate the recall (sensitivity) of predictions}
\usage{
recall(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
recall value
}
\description{
Calculate the recall (sensitivity) of predictions
}
\examples{
# Define predicted and true values
pred <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
truth <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Calculate recall
recall(pred, truth)

}
