% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{rep_mat}
\alias{rep_mat}
\title{Replicate a matrix or dataframe M times vertically and N times horizontally}
\usage{
rep_mat(x, M = 1, N = 1)
}
\arguments{
\item{x}{a dataframe or matrix}

\item{M}{number of times to replicate vertically (default = 1)}

\item{N}{number of times to replicate horizontally (default = 1)}
}
\value{
a replicated matrix or dataframe
}
\description{
Replicate a matrix or dataframe M times vertically and N times horizontally
}
\examples{
> mat <- matrix(1:4, nrow = 2, ncol = 2)
> mat
[,1] [,2]
[1,]    1    3
[2,]    2    4
> ( mat <- matrix(1:4, nrow = 2, ncol = 2))
[,1] [,2]
[1,]    1    3
[2,]    2    4
> (replicated_mat <- rep_mat(mat, M = 2, N = 3))
[,1] [,2] [,3] [,4] [,5] [,6]
[1,]    1    3    1    3    1    3
[2,]    2    4    2    4    2    4
[3,]    1    3    1    3    1    3
[4,]    2    4    2    4    2    4

}
