% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{sensitivity}
\alias{sensitivity}
\title{Calculate the sensitivity (true positive rate) of predictions}
\usage{
sensitivity(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
sensitivity value
}
\description{
Calculate the sensitivity (true positive rate) of predictions
}
\examples{
# Define predicted and true values
pred <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
truth <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Calculate sensitivity
sensitivity(pred, truth)

}
