% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi510.R
\name{specificity}
\alias{specificity}
\title{Calculate the specificity (true negative rate) of predictions}
\usage{
specificity(pred, truth)
}
\arguments{
\item{pred}{predicted values}

\item{truth}{true values}
}
\value{
specificity value
}
\description{
Calculate the specificity (true negative rate) of predictions
}
\examples{
# Define predicted and true values
pred <- c(TRUE, FALSE, TRUE, TRUE, FALSE)
truth <- c(TRUE, FALSE, TRUE, FALSE, TRUE)

# Calculate specificity
specificity(pred, truth)

}
